CH: restructure directory layout
